+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Able to setup context without actions
+
 ## [1.0.0] - 2025-03-23
 
 - Initial release